Recipe ML model/requirements.txt
    absl-py==2.1.0
    appdirs==1.4.4
    blinker==1.7.0
    boto3==1.34.72
    botocore==1.34.72
    build==0.10.0
    certifi==2024.2.2
    charset-normalizer==3.3.2
    chex==0.1.85
    click==8.1.7
    colorama==0.4.6
    etils==1.7.0
    filelock==3.13.1
    Flask==3.0.2
    Flask-Cors==4.0.0
    flax==0.8.1
    fsspec==2024.2.0
    gunicorn==21.2.0
    huggingface-hub==0.21.4
    idna==3.6
    importlib_metadata==7.1.0
    importlib_resources==6.3.0
    itsdangerous==2.1.2
    jax==0.4.25
    jaxlib==0.4.25
    Jinja2==3.1.3
    jmespath==1.0.1
    markdown-it-py==3.0.0
    MarkupSafe==2.1.5
    mdurl==0.1.2
    ml-dtypes==0.3.2
    modelbit==0.35.0
    msgpack==1.0.8
    nest-asyncio==1.6.0
    numpy==1.26.4
    opt-einsum==3.3.0
    optax==0.2.1
    orbax-checkpoint==0.5.5
    packaging==24.0
    pandas==2.2.1
    pkginfo==1.10.0
    protobuf==4.25.3
    pycryptodomex==3.20.0
    Pygments==2.17.2
    pyproject_hooks==1.0.0
    python-dateutil==2.9.0.post0
    python-dotenv==1.0.1
    pytz==2024.1
    PyYAML==6.0.1
    regex==2023.12.25
    requests==2.31.0
    rich==13.7.1
    s3transfer==0.10.1
    safetensors==0.4.2
    scipy==1.12.0
    six==1.16.0
    tensorstore==0.1.54
    texttable==1.7.0
    tokenizers==0.15.2
    toolz==0.12.1
    tqdm==4.66.2
    transformers==4.38.2
    types-pkg-resources==0.1.3
    types-PyYAML==6.0.12.20240311
    types-requests==2.31.0.6
    types-urllib3==1.26.25.14
    typing_extensions==4.10.0
    tzdata==2024.1
    urllib3==1.26.18
    Werkzeug==3.0.1
    zipp==3.18.0
    zstandard==0.22.0
